pyproject.toml
[tool.poetry]
name = "lhy"
version = "0.1.0"
description = ""
authors = ["이한영 <dev@lhy.kr>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
drf-yasg = "^1.17"
pillow = "^6.2"
django-extensions = "^2.2"
django-filter = "^2.2"
django-admin-sortable2 = "^0.7.3"
django-cleanup = "^4.0"
django-phonenumber-field = "^3.0"
django-storages = "^1.7"
easy-thumbnails = "^2.6"
djangorestframework-camel-case = "^1.1"
django-cors-headers = "^3.1"
django-rest-auth = "^0.9.5"
drf-writable-nested = "^0.5.1"
psycopg2-binary = "^2.8"
jinja2 = "^2.10"
phonenumberslite = "^8.10"
django-aws-secrets-manager = "^0.1.2"
django-modeladmin-reorder = "^0.3.1"
boto3 = "^1.10"
django-debug-toolbar = "^2.0"
django-default-imagefield = "^0.5.5"
gunicorn = "^19.9"
django-dbbackup = "<3.2"
django-allauth = "^0.40.0"
subdomains = "^2.1.1"
django-polymorphic = "^2.1.2"

[tool.poetry.dev-dependencies]
notebook = "^6.0.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
